moving font registration to registrar class
